Rename Trustable to GrimoireLab Metrics
Rename Trustable to GrimoireLab Metrics for consistency with the current code.

This commit also includes a schema of how it works.
