+++ tests/end_to_end/test_cli.py
@@ -20,21 +20,21 @@ import json
 import logging
 import unittest
 
-from trustable_cli.cli import trustable_grimoirelab_score
+from grimoirelab_metrics.cli import grimoirelab_metrics
 from end_to_end.base import EndToEndTestCase
 
 GRIMOIRELAB_URL = "http://localhost:8000"
 
 
 class TestMetrics(EndToEndTestCase):
-    """End to end tests for Trustable CLI metrics"""
+    """End to end tests for grimoirelab metrics CLI"""
 
     def test_metrics(self):
         """Check whether the metrics are correctly calculated"""
 
         with self.assertLogs(logging.getLogger()) as logger:
             result = self.runner.invoke(
-                trustable_grimoirelab_score,
+                grimoirelab_metrics,
                 [
                     "./data/archived_repos.spdx.xml",
                     "--grimoirelab-url",
@@ -118,7 +118,7 @@ class TestMetrics(EndToEndTestCase):
 
         with self.assertLogs(logging.getLogger()) as logger:
             result = self.runner.invoke(
-                trustable_grimoirelab_score,
+                grimoirelab_metrics,
                 [
                     "./data/archived_repos.spdx.xml",
                     "--grimoirelab-url",
@@ -202,7 +202,7 @@ class TestMetrics(EndToEndTestCase):
 
         with self.assertLogs(logging.getLogger()) as logger:
             result = self.runner.invoke(
-                trustable_grimoirelab_score,
+                grimoirelab_metrics,
                 [
                     "./data/archived_repos.spdx.xml",
                     "--grimoirelab-url",
@@ -286,7 +286,7 @@ class TestMetrics(EndToEndTestCase):
 
         with self.assertLogs(logging.getLogger()) as logger:
             result = self.runner.invoke(
-                trustable_grimoirelab_score,
+                grimoirelab_metrics,
                 [
                     "./data/duplicate_repo.spdx.xml",
                     "--grimoirelab-url",
@@ -346,7 +346,7 @@ class TestMetrics(EndToEndTestCase):
 
         with self.assertLogs(logging.getLogger()) as logger:
             result = self.runner.invoke(
-                trustable_grimoirelab_score,
+                grimoirelab_metrics,
                 [
                     "./data/mercurial_repo.spdx.xml",
                     "--grimoirelab-url",
